Move match function to helpers
Update match function to operate on familiar names
