+++ reference/reference.go
@@ -27,7 +27,6 @@ package reference
 import (
 	"errors"
 	"fmt"
-	"path"
 	"strings"
 
 	"github.com/opencontainers/go-digest"
@@ -315,16 +314,6 @@ func WithDigest(name Named, digest digest.Digest) (Canonical, error) {
 		namedRepository: repo,
 		digest:          digest,
 	}, nil
-}
-
-// Match reports whether ref matches the specified pattern.
-// See https://godoc.org/path#Match for supported patterns.
-func Match(pattern string, ref Reference) (bool, error) {
-	matched, err := path.Match(pattern, ref.String())
-	if namedRef, isNamed := ref.(Named); isNamed && !matched {
-		matched, _ = path.Match(pattern, namedRef.Name())
-	}
-	return matched, err
 }
 
 // TrimNamed removes any tag or digest from the named reference.
